Need to import db different way

+++ src/main/java/com/musicfeels/MusicFeels/repository/ChordProgressionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChordProgressionRepository extends CrudRepository<ChordProgression, Integer> {
+public interface ChordProgressionRepository extends CrudRepository<ChordProgression, Long> {
 
 }
